- Trying to fix failed push

+++ Blaze/TextFields/BlazePickerViewField.m
@@ -88,7 +88,7 @@
 -(NSString *)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component
 {
     if(row >= self.pickerValues.count) {
-        NSLog(@"WARNING - Row index %d too high for pickervalues count %d, you should check this! Returning nil for now...", row, self.pickerValues.count);
+        NSLog(@"WARNING - Row index %ld too high for pickervalues count %ld, you should check this! Returning nil for now...", (long)row, self.pickerValues.count);
         return nil;
     }
     return self.pickerValues[row];
